requirements.txt
langchain
langchain-community
openai
chromadb
tiktoken
python-dotenv